benchmarks/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
anykeystore==0.2
apex @ file:///home/vedaant/EVA/EVA-CLIP/rei/apex
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
cmake==3.27.0
contourpy==1.1.0
cryptacular==1.6.2
cycler==0.11.0
datasets==2.14.0
deepspeed==0.10.0
defusedxml==0.7.1
dill==0.3.7
docker-pycreds==0.4.0
einops==0.6.1
filelock==3.12.2
fonttools==4.41.1
frozenlist==1.4.0
fsspec==2023.6.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.2
hjson==3.1.0
huggingface-hub==0.16.4
hupper==1.12
idna==3.4
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
lit==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.1
ninja==1.11.1
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.74
packaging==23.1
pandas==2.0.3
PasteDeploy==3.0.1
pathtools==0.1.2
pbkdf2==1.3
Pillow==10.0.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
protobuf==4.23.4
psutil==5.9.5
py-cpuinfo==9.0.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.12
pyparsing==3.0.9
pyramid==2.0.1
pyramid-mailer==0.15.1
pyre-extensions==0.0.29
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.6.3
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.19
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.1
threadpoolctl==3.2.0
timm==0.9.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transaction==3.1.0
transformers @ git+https://github.com/huggingface/transformers@e7b001db4fbd33d77de95cf684d13d7605660d1b
translationstring==1.4
triton==2.0.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
velruse==1.1.1
venusian==3.0.0
wandb==0.15.7
wcwidth==0.2.6
WebOb==1.8.7
Werkzeug==2.3.6
WTForms==3.0.1
wtforms-recaptcha==0.3.2
xformers==0.0.20
xxhash==3.2.0
yarl==1.9.2
zope.deprecation==5.0
zope.interface==6.0
zope.sqlalchemy==3.0
